style: hook 2개 한개로 수정
