+++ src/pages/home/subPages/components/memo/ImageMemoText/index.ts
@@ -1,1 +1,2 @@
 export { default as ImageMemoText } from './ImageMemoText';
+export * from './MemoText';
