+++ src/pages/home/subPages/components/memo/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useMemoManager } from './useMemoManager';
