--- src/pages/home/subPages/components/memo/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useMemoDeleteManager } from './useMemoDeleteManager';
-export { default as useMemoUpdateManager } from './useMemoUpdateManager';
