--- src/pages/home/subPages/components/memo/hooks/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './infiniteQuery';
