--- src/pages/home/subPages/components/memo/hooks/interfaces/infiniteQuery.ts
@@ -1,12 +1,0 @@
-import { Memo } from 'pages/home/subPages/interfaces';
-
-export interface InfiniteQueryData {
-  pages: MemoData[];
-  pageParams: number[];
-}
-
-export interface MemoData {
-  current_count: number;
-  current_page: number;
-  memos: Memo[];
-}
