+++ src/pages/home/subPages/components/memo/index.ts
@@ -2,3 +2,4 @@ export * from './EditableMemo';
 export * from './MemosList';
 export * from './UneditableMemo';
 export * from './ImageMemoText';
+export * from './hook';
